WIP: A try at constant FPS with Delay sypport for the game loop.
Needs more test and advice.

After some tests, it seems the game update and render logic must be called
the same times in a loop cycle to let the game be as smooth as possible.
Thus, I didn't separate the game update speed from the frame per seconds logic.

Questions:
- Should we keep the SDL_Delay at all?
- Should we add an option to let the user choose the update method agressiveness?
 - Gentle: Use delay.
 - Capped: Render at 60 FPS.
 - Agressive: Not capping or delay, we update as much as possible.